Moved tokens to general namespace
